Changes required to enable sphtorus to run MPI.
Communication of particle data is performed by a call to the new subroutine gather_sph_data, following a call to new subroutine init_sph_data2 to set up the sph data structure. The aim was to encapsulate the setting up of the data structure and the PI communication within separate subroutines rather than adding new code to torusMain.

The Makefile has been modified to pre-process sph_data_class.f90. This is not the ideal way to do the pre-processing (better to change file extention to .F90) so should be changed later.


git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@1060 25464f21-6099-402c-b30a-640f1d40d38b