Added two new parameters to the input variables read from the parameters.dat file.
The first is an integer parameter called iterLucy which sets the minimum number of iterations of the Lucy algorithm. This replaces the hardwired value in torusMain. The default value is 10.

The second new parameter is a logical called forceLucyConv. This allows convergence of the Lucy algorithm to be forced after a single iteration. This change uses code which was already in lucy_mod (but commented out). The default is .false.

Using parameters read from in parameters.dat allows changes to be made without having to recompile the code.


git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@1163 25464f21-6099-402c-b30a-640f1d40d38b