Fix to prevent crash in MPI benchmark runs. The resulting 2D vtk file cannot be read by visit but the code no longer crashes.
If the grid is not split over MPI then the VTK call is only made by process zero (called from do_amr_plots in TorusMain). In this case the header needs to be written even if the global rank is not 1.

The writeHeader logical has been put into a module variable so that it can be consistently set across all parts of vtk_mod. The call to writeVtkFile sets the value of writeHeader. Subroutines have been declared public or private in order to prevent calls to subroutines when writeHeader is not set i.e. need to ensure that entry to vtk_mod is via writeVtkFile so that writeHeader is set correctly.


git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@1357 25464f21-6099-402c-b30a-640f1d40d38b