First step of the work to improve image generation.
Various parameters in input_variable have been removed, restructured or moved. The aim is to clarify what already exists in terms of data cube and image 
generation. 

Some input parameters used in phaseloop_mod are not being set. These have been moved into phaseloop_mode and will either be deleted (if the 
corresponding code is not required) or re-instated as an input variable once the corresponding code is fixed. 




git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@2984 25464f21-6099-402c-b30a-640f1d40d38b