Changes for Galactic CO cubes.
Allow a temperature to be used for setting the maximum level of interest in molecular_mod
Add an input parameter to indicate that SPH chemistry data need to be loaded (sphwithchem)
Change to inputs_mod to read SPH chemistry parameters when h21cm=.false.
Added some extra output to clarify what is happening in the SPH to grid conversion.



git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@3620 25464f21-6099-402c-b30a-640f1d40d38b