Changes for CO Galactic plane surveys.
1. Store total density in the particle rho component. This means that grid splitting is always done using total 
rho and never HI rho. The latter will give poor resolution in regions with low HI density where CO density should
be high.

2. Control conversion of density to HI density using only the convertRhoToHI flag i.e. allow chemistry runs to store
total density on the grid. This allows density sub-sampling to be used with CO cubes.

3. The H2 fraction written when mapping to particles in angularImage_mod is now a fraction by mass which is 
 consistent whether or not convertRhoToHI is in use. 

4. There is a new logical in the parameters list which causes molecular loop to set LTE level populations then 
exit. This is required for making LTE Galactic plane surveys. This logical is setupMolecularLteOnly. 



git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@4209 25464f21-6099-402c-b30a-640f1d40d38b