Update SPH to grid conversion. The aim is to make the code more maintainable (including making it easier to grid new SPH variables) and to add some parallelism using OpenMP.
The clusterParameter function has been modified to take optional arguments and has been restructured to make it clearer how to add new 
variables. 

Most of the SAVEd variables have been removed to help add OpenMP. If the gridding is to run in parallel the algorithm should not depend on 
the state from a previous call. This required restructing the calls to clusterParameter so that it is only called once per grid point 
(instead of relying on the reuse behaviour which required saved variables). 

Gridding with OpenMP works correctly with the gfortran compiler but generates a seg fault with ifort. These seg fault occurs in 
clusterParameter the first time it makes a call to a sbroutine or function so I suspect this is a stack problem. As parallel gridding 
does not always work this is left as a switchable option at build time (include with sph2grid_parallel=yes on the make line). 



git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@5189 25464f21-6099-402c-b30a-640f1d40d38b